backend/requirements.txt
# requirements.txt — Python backend
# Pin all versions for reproducibility

fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
tensorflow==2.16.1
python-dotenv==1.0.1
supabase==2.4.2
apscheduler==3.10.4
pydantic==2.7.1
python-multipart==0.0.9
